requirements.txt
streamlit>=1.28.1
openai>=1.107.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
